master | hot-fix add bloock provider and persist token when do operations to comunicate with wallet

+++ internal/services/create/request/issuer_request.go
@@ -6,7 +6,7 @@ type CreateIssuerRequest struct {
 	Name            string
 	Description     string
 	Image           string
-	PublishInterval int64
+	PublishInterval int
 }
 
 type DidMetadataRequest struct {
